feat: updated types for api and env
