+++ src/commands/base-command.ts
@@ -1,5 +1,3 @@
-import { isCI } from 'ci-info'
-
 import { existsSync } from 'fs'
 import { join, relative, resolve } from 'path'
 import process from 'process'
@@ -8,6 +6,7 @@ import { format } from 'util'
 import { DefaultLogger, Project } from '@netlify/build-info'
 import { NodeFS, NoopLogger } from '@netlify/build-info/node'
 import { resolveConfig } from '@netlify/config'
+import { isCI } from 'ci-info'
 import { Command, Help, Option } from 'commander'
 // @ts-expect-error TS(7016) FIXME: Could not find a declaration file for module 'debu... Remove this comment to see the full error message
 import debug from 'debug'
