+++ src/utils/types.ts
@@ -62,3 +62,7 @@ export interface Request extends IncomingMessage {
 }
 
 export type Rewriter = (req: Request) => Match | null
+
+export interface APIEnvError {
+  json: { msg: string }
+}
